restructure cli.py->run.py, main.py->app.py (#141)
because one can simply launch factgenie/run.py
and run is entry point for cli launch, create_app factory used for click factgenie command
click CLI commands
and the app module really just defines the app and is not entry point at all
